refactor (#450)
