+++ git-refspec/src/lib.rs
@@ -33,6 +33,7 @@ mod write;
 
 ///
 pub mod match_group;
+pub use match_group::types::MatchGroup;
 
 mod types;
-pub use types::{Instruction, MatchGroup};
+pub use types::Instruction;
